+++ internal/ui/help/model.go
@@ -102,6 +102,7 @@ func (m Model) View() string {
 				{"1-9", styles.IconView + "Switch views (tabs)"},
 				{"f", styles.IconTag + "Filter by tag (Tag view)"},
 				{getK(m.km.OpenTask), styles.IconView + "View task details"},
+				{getK(m.km.ProjectSwitcher), "󰜎 " + "Switch project"},
 				{getK(m.km.Back), "󰌍 " + "Back / Close"},
 			},
 		},